+++ apps/zoos/src/routeTree.gen.ts
@@ -39,7 +39,7 @@ import { Route as CoreReactTableMergeColumnsImport } from './routes/core/react-t
 import { Route as CoreReactTableFullCustomizationImport } from './routes/core/react-table/full-customization'
 import { Route as CoreReactTableExpandRowsControlImport } from './routes/core/react-table/expand-rows-control'
 import { Route as CoreReactTableExpandRowsCellImport } from './routes/core/react-table/expand-rows-cell'
-import { Route as CoreReactTableCustomColumnsImport } from './routes/core/react-table/custom-columns'
+import { Route as CoreReactTableColumnFiltersImport } from './routes/core/react-table/column-filters'
 import { Route as CoreReactQueryUseQueryImport } from './routes/core/react-query/use-query'
 import { Route as CoreReactQueryUseMutationImport } from './routes/core/react-query/use-mutation'
 import { Route as CoreReactFormWithZoosFormImport } from './routes/core/react-form/with-zoos-form'
@@ -231,10 +231,10 @@ const CoreReactTableExpandRowsCellRoute =
     getParentRoute: () => CoreReactTableRouteRoute,
   } as any)
 
-const CoreReactTableCustomColumnsRoute =
-  CoreReactTableCustomColumnsImport.update({
-    id: '/custom-columns',
-    path: '/custom-columns',
+const CoreReactTableColumnFiltersRoute =
+  CoreReactTableColumnFiltersImport.update({
+    id: '/column-filters',
+    path: '/column-filters',
     getParentRoute: () => CoreReactTableRouteRoute,
   } as any)
 
@@ -485,11 +485,11 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof CoreReactQueryUseQueryImport
       parentRoute: typeof CoreReactQueryRouteImport
     }
-    '/core/react-table/custom-columns': {
-      id: '/core/react-table/custom-columns'
-      path: '/custom-columns'
-      fullPath: '/core/react-table/custom-columns'
-      preLoaderRoute: typeof CoreReactTableCustomColumnsImport
+    '/core/react-table/column-filters': {
+      id: '/core/react-table/column-filters'
+      path: '/column-filters'
+      fullPath: '/core/react-table/column-filters'
+      preLoaderRoute: typeof CoreReactTableColumnFiltersImport
       parentRoute: typeof CoreReactTableRouteImport
     }
     '/core/react-table/expand-rows-cell': {
@@ -654,7 +654,7 @@ const CoreReactQueryRouteRouteWithChildren =
   CoreReactQueryRouteRoute._addFileChildren(CoreReactQueryRouteRouteChildren)
 
 interface CoreReactTableRouteRouteChildren {
-  CoreReactTableCustomColumnsRoute: typeof CoreReactTableCustomColumnsRoute
+  CoreReactTableColumnFiltersRoute: typeof CoreReactTableColumnFiltersRoute
   CoreReactTableExpandRowsCellRoute: typeof CoreReactTableExpandRowsCellRoute
   CoreReactTableExpandRowsControlRoute: typeof CoreReactTableExpandRowsControlRoute
   CoreReactTableFullCustomizationRoute: typeof CoreReactTableFullCustomizationRoute
@@ -663,7 +663,7 @@ interface CoreReactTableRouteRouteChildren {
 }
 
 const CoreReactTableRouteRouteChildren: CoreReactTableRouteRouteChildren = {
-  CoreReactTableCustomColumnsRoute: CoreReactTableCustomColumnsRoute,
+  CoreReactTableColumnFiltersRoute: CoreReactTableColumnFiltersRoute,
   CoreReactTableExpandRowsCellRoute: CoreReactTableExpandRowsCellRoute,
   CoreReactTableExpandRowsControlRoute: CoreReactTableExpandRowsControlRoute,
   CoreReactTableFullCustomizationRoute: CoreReactTableFullCustomizationRoute,
@@ -761,7 +761,7 @@ export interface FileRoutesByFullPath {
   '/core/react-form/with-zoos-form': typeof CoreReactFormWithZoosFormRoute
   '/core/react-query/use-mutation': typeof CoreReactQueryUseMutationRoute
   '/core/react-query/use-query': typeof CoreReactQueryUseQueryRoute
-  '/core/react-table/custom-columns': typeof CoreReactTableCustomColumnsRoute
+  '/core/react-table/column-filters': typeof CoreReactTableColumnFiltersRoute
   '/core/react-table/expand-rows-cell': typeof CoreReactTableExpandRowsCellRoute
   '/core/react-table/expand-rows-control': typeof CoreReactTableExpandRowsControlRoute
   '/core/react-table/full-customization': typeof CoreReactTableFullCustomizationRoute
@@ -803,7 +803,7 @@ export interface FileRoutesByTo {
   '/core/react-form/with-zoos-form': typeof CoreReactFormWithZoosFormRoute
   '/core/react-query/use-mutation': typeof CoreReactQueryUseMutationRoute
   '/core/react-query/use-query': typeof CoreReactQueryUseQueryRoute
-  '/core/react-table/custom-columns': typeof CoreReactTableCustomColumnsRoute
+  '/core/react-table/column-filters': typeof CoreReactTableColumnFiltersRoute
   '/core/react-table/expand-rows-cell': typeof CoreReactTableExpandRowsCellRoute
   '/core/react-table/expand-rows-control': typeof CoreReactTableExpandRowsControlRoute
   '/core/react-table/full-customization': typeof CoreReactTableFullCustomizationRoute
@@ -848,7 +848,7 @@ export interface FileRoutesById {
   '/core/react-form/with-zoos-form': typeof CoreReactFormWithZoosFormRoute
   '/core/react-query/use-mutation': typeof CoreReactQueryUseMutationRoute
   '/core/react-query/use-query': typeof CoreReactQueryUseQueryRoute
-  '/core/react-table/custom-columns': typeof CoreReactTableCustomColumnsRoute
+  '/core/react-table/column-filters': typeof CoreReactTableColumnFiltersRoute
   '/core/react-table/expand-rows-cell': typeof CoreReactTableExpandRowsCellRoute
   '/core/react-table/expand-rows-control': typeof CoreReactTableExpandRowsControlRoute
   '/core/react-table/full-customization': typeof CoreReactTableFullCustomizationRoute
@@ -894,7 +894,7 @@ export interface FileRouteTypes {
     | '/core/react-form/with-zoos-form'
     | '/core/react-query/use-mutation'
     | '/core/react-query/use-query'
-    | '/core/react-table/custom-columns'
+    | '/core/react-table/column-filters'
     | '/core/react-table/expand-rows-cell'
     | '/core/react-table/expand-rows-control'
     | '/core/react-table/full-customization'
@@ -935,7 +935,7 @@ export interface FileRouteTypes {
     | '/core/react-form/with-zoos-form'
     | '/core/react-query/use-mutation'
     | '/core/react-query/use-query'
-    | '/core/react-table/custom-columns'
+    | '/core/react-table/column-filters'
     | '/core/react-table/expand-rows-cell'
     | '/core/react-table/expand-rows-control'
     | '/core/react-table/full-customization'
@@ -978,7 +978,7 @@ export interface FileRouteTypes {
     | '/core/react-form/with-zoos-form'
     | '/core/react-query/use-mutation'
     | '/core/react-query/use-query'
-    | '/core/react-table/custom-columns'
+    | '/core/react-table/column-filters'
     | '/core/react-table/expand-rows-cell'
     | '/core/react-table/expand-rows-control'
     | '/core/react-table/full-customization'
@@ -1090,7 +1090,7 @@ export const routeTree = rootRoute
     "/core/react-table": {
       "filePath": "core/react-table/route.tsx",
       "children": [
-        "/core/react-table/custom-columns",
+        "/core/react-table/column-filters",
         "/core/react-table/expand-rows-cell",
         "/core/react-table/expand-rows-control",
         "/core/react-table/full-customization",
@@ -1180,8 +1180,8 @@ export const routeTree = rootRoute
       "filePath": "core/react-query/use-query.tsx",
       "parent": "/core/react-query"
     },
-    "/core/react-table/custom-columns": {
-      "filePath": "core/react-table/custom-columns.tsx",
+    "/core/react-table/column-filters": {
+      "filePath": "core/react-table/column-filters.tsx",
       "parent": "/core/react-table"
     },
     "/core/react-table/expand-rows-cell": {