+++ packages/react-table-ui/src/header/header-context-menu.tsx
@@ -30,7 +30,9 @@ const HeaderContextMenu = <TData, TValue>({
 
   return (
     <>
-      <ContextMenuTrigger {...rest}>{children}</ContextMenuTrigger>
+      <ContextMenuTrigger {...rest}>
+        <div>{children}</div>
+      </ContextMenuTrigger>
       <ContextMenuContent className="w-48">
         <ContextMenuLabel>{children}</ContextMenuLabel>
         <ContextMenuSeparator className="mx-1 border-b" />